+++ test/test_rename.rb
@@ -5,9 +5,10 @@ commandparameter = " rename -o #{TMPDIR} -c false "
 `#{PDFMD} #{commandparameter} #{TARGETPDF}`.chomp
 files = readFilesInDir(TMPDIR)
 if files.size == 1 and
-  File.basename(files.keys[0]) == '19700101-example_author-dok-cno1111111-kdn1111111-test_subject.pdf'
+  File.basename(files.keys[0]) == '19700101-example_author-doc-test_subject-cno1111111-some_keywords.pdf'
   result = 'OK'
 else
+  puts File.basename(files.keys[0])
   result = 'failed'
 end
 $testResults = { '001' => {:result => result, :command => commandparameter }}
@@ -20,10 +21,11 @@ commandparameter = " rename -c -o #{TMPDIR}"
 `#{PDFMD} #{commandparameter} #{TARGETPDF}`.chomp
 files = readFilesInDir(TMPDIR)
 if files.size == 2 and
-  File.basename(files.keys[0]) == '19700101-example_author-dok-cno1111111-kdn1111111-test_subject.pdf' and
+  File.basename(files.keys[0]) == '19700101-example_author-doc-test_subject-cno1111111-some_keywords.pdf'
   File.basename(files.keys[1]) == 'test_default.pdf'
   result = 'OK'
 else
+  puts File.basename(files.keys[0])
   result = 'failed'
 end
 $testResults.store('002', {:result => result, :command => commandparameter })
@@ -49,9 +51,10 @@ commandparameter = " rename -a -o #{TMPDIR} -c false"
 `#{PDFMD} #{commandparameter} #{TARGETPDF}`.chomp
 files = readFilesInDir(TMPDIR)
 if files.size == 1 and
-  File.basename(files.keys[0]) == '19700101-example_author-dok-cno1111111-kdn1111111-test_subject-some_keywords-author-some_feature.pdf'
+  File.basename(files.keys[0]) == '19700101-example_author-doc-test_subject-cno1111111-some_keywords-author-some_feature-kundenummer_1111111.pdf'
   result = 'OK'
 else
+  puts File.basename(files.keys[0])
   result = 'failed'
 end
 $testResults.store('004', {:result => result, :command => commandparameter })
@@ -64,9 +67,10 @@ commandparameter = " rename -k 1 -o #{TMPDIR} -c false"
 `#{PDFMD} #{commandparameter} #{TARGETPDF}`.chomp
 files = readFilesInDir(TMPDIR)
 if files.size == 1 and
-  File.basename(files.keys[0]) == '19700101-example_author-dok-test_subject.pdf'
+  File.basename(files.keys[0]) == '19700101-example_author-doc-test_subject.pdf'
   result = 'OK'
 else
+  puts File.basename(files.keys[0])
   result = 'failed'
 end
 $testResults.store('005', {:result => result, :command => commandparameter })
@@ -78,10 +82,11 @@ commandparameter = " rename -l -p #{TMPDIR}/pdfmd.log -c false"
 `#{PDFMD} #{commandparameter} #{TARGETPDF}`.chomp
 files = readFilesInDir(TMPDIR)
 if files.size == 1 and
-  File.basename(files.keys[0]) == '19700101-example_author-dok-cno1111111-kdn1111111-test_subject.pdf' and
+  File.basename(files.keys[0]) == '19700101-example_author-doc-test_subject-cno1111111-some_keywords.pdf' and
   `tail -n 1 #{TMPDIR}/pdfmd.log | wc -l`.to_i == 1
   result = 'OK'
 else
+  puts File.basename(files.keys[0])
   result = 'failed'
 end
 $testResults.store('006', {:result => result, :command => commandparameter })